fix typeerrors n stuff

+++ src/components/Samples/Samples.tsx
@@ -1,51 +1,57 @@
 import { useEffect, useRef, useState } from 'react';
 
 import {
-  createSample,
+  createSampleRecord,
   fetchSamples,
   deleteSample,
   getPocketBase,
-} from '../../db/db_samples';
+} from '../../db/samplesPb';
 import { fetchBlobFromUrl } from '../../utils/fetch';
 import { Sample } from '../../types';
 import styles from './Samples.module.scss';
 
 interface SamplesProps {
   currentSampleUrl: string;
-  // currentSampleBlob: Blob; // henda eftir test
-  chooseSample: (audioUrl: string) => void;
+  handleChooseSample: (audioUrl: string) => void;
 }
 
 const Samples: React.FC<SamplesProps> = ({
   currentSampleUrl,
-  // currentSampleBlob,
-  chooseSample,
+  handleChooseSample,
 }) => {
   const pocketBase = getPocketBase();
 
+  const [currentSampleSrc, setCurrentSampleSrc] = useState<string>('');
   const [userSamples, setUserSamples] = useState<Sample[]>([]);
+  const [showSampleList, setShowSampleList] = useState<boolean>(false);
 
   async function loadUserSamples() {
     const sampleObjArray = await fetchSamples();
     setUserSamples(sampleObjArray);
   }
 
-  const [showSampleList, setShowSampleList] = useState<boolean>(false);
+  useEffect(() => {
+    setCurrentSampleSrc(currentSampleUrl); // Set currentSampleSrc whenever currentSampleUrl changes
+    console.log('currentSampleUrl: ', currentSampleUrl);
+  }, [currentSampleUrl]);
 
-  const toggleShowSamples = (): void => {
-    loadUserSamples();
+  const toggleShowSamples = async (): Promise<void> => {
     setShowSampleList((prevShowSampleList) => !prevShowSampleList);
+    if (!showSampleList) {
+      await loadUserSamples();
+    }
   };
 
   const handleSave = async (): Promise<void> => {
     const name = prompt('Enter a name for the sample:');
     if (name !== null && name.trim() !== '') {
-      // if (currentSampleBlob) {
-      //   await createSample(name, currentSampleBlob);
-      // } else
       if (currentSampleUrl) {
+        // currentSampleUrl ?
         const newBlob = await fetchBlobFromUrl(currentSampleUrl);
-        await createSample(name, newBlob);
+        await createSampleRecord(name, newBlob);
+      } else {
+        console.error('No sample to save.');
+        alert('No sample to save.');
       }
       await loadUserSamples();
     } else {
@@ -54,15 +60,9 @@ const Samples: React.FC<SamplesProps> = ({
     }
   };
 
-  const confirmDelete = (sample: Sample) => {
-    const isConfirmed = window.confirm('Delete this sample?');
-    if (isConfirmed) {
-      handleDelete(sample);
-    }
-  };
-
-  async function handleDelete(sample: Sample): Promise<void> {
-    if (sample.id) {
+  const handleDelete = async (sample: Sample): Promise<void> => {
+    const isConfirmed = window.confirm('Delete this sample?'); // create custom
+    if (isConfirmed && sample.id) {
       try {
         await deleteSample(sample.id);
         await loadUserSamples();
@@ -71,13 +71,13 @@ const Samples: React.FC<SamplesProps> = ({
         alert('Delete sample failed');
       }
     }
-  }
+  };
 
   const downloadAudio = () => {
     // nota pocketbase frekar
-    if (currentSampleUrl) {
+    if (currentSampleSrc) {
       const link = document.createElement('a');
-      link.href = currentSampleUrl;
+      link.href = currentSampleSrc;
       link.download = 'sample-hljodsmali.wav';
       document.body.appendChild(link);
       link.click();
@@ -96,31 +96,36 @@ const Samples: React.FC<SamplesProps> = ({
       </div>
       {showSampleList && (
         <ul
-          className={`${styles.samplesList} ${
-            showSampleList ? styles.visible : ''
-          }`}
+          className={`${styles.samplesList} 
+          ${showSampleList ? styles.visible : ''}
+          `}
         >
-          {userSamples.map((sample, index) => (
-            <li key={index}>
-              <button
-                onClick={() => chooseSample(sample.audioUrl!)}
-                className={styles.singleSample}
-                data-src={sample.audioUrl}
-              >
-                {sample.name}
-              </button>
+          {userSamples.map(
+            (
+              sample,
+              index // use index?
+            ) => (
+              <li key={sample.id}>
+                <button
+                  onClick={() => handleChooseSample(sample.audioUrl!)}
+                  className={styles.singleSample}
+                  data-src={sample.audioUrl}
+                >
+                  {sample.name}
+                </button>
 
-              {/* <audio src={sample.audioUrl}></audio> */}
+                {/* <audio src={sample.audioUrl}></audio> */}
 
-              <button
-                onClick={() => confirmDelete(sample)}
-                id={sample.id}
-                className={styles.deleteButton}
-              >
-                x
-              </button>
-            </li>
-          ))}
+                <button
+                  onClick={() => handleDelete(sample)}
+                  id={sample.id}
+                  className={styles.deleteButton}
+                >
+                  x
+                </button>
+              </li>
+            )
+          )}
         </ul>
       )}
     </div>
